backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
av==13.1.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
cloudinary==1.44.0
cloup==3.0.7
colorama==0.4.6
decorator==5.2.1
Flask==3.1.1
flask-cors==5.0.1
glcontext==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-genai==1.16.1
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
isosurfaces==0.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
manim==0.19.0
ManimPango==0.6.0
mapbox_earcut==1.0.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
moderngl==5.12.0
moderngl-window==3.1.1
networkx==3.4.2
numpy==2.2.6
pillow==11.2.1
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.28.0
pydantic==2.11.4
pydantic_core==2.33.2
pydub==0.25.1
pyglet==2.1.6
pyglm==2.8.2
Pygments==2.19.1
pyparsing==3.2.3
python-dotenv==1.1.0
requests==2.32.3
rich==14.0.0
rsa==4.9.1
scipy==1.15.3
screeninfo==0.8.1
six==1.17.0
skia-pathops==0.8.0.post2
sniffio==1.3.1
soupsieve==2.7
srt==3.5.3
svgelements==1.9.6
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
watchdog==6.0.0
websockets==15.0.1
Werkzeug==3.1.3
threading